Cast to avoid overflow

+++ src/graphview/extendedgraphview.cpp
@@ -63,7 +63,7 @@ void ExtendedGraphView::rescalePlot()
     else if (_pGuiModel->xAxisScalingMode() == SCALE_SLIDING)
     {
         // sliding window scale routine
-        const quint64 slidingInterval = _pGuiModel->xAxisSlidingSec() * 1000;
+        const quint64 slidingInterval = static_cast<quint64>(_pGuiModel->xAxisSlidingSec()) * 1000;
         if ((_pPlot->graphCount() != 0) && (graphDataSize() != 0))
         {
             auto lastDataIt = _pPlot->graph(0)->data()->constEnd();
